+++ packages/gov-ui-kit/src/modules/components/proposal/proposalActions/proposalActionsList/proposalActionWithdrawToken/proposalActionWithdrawToken.api.tsx
@@ -15,6 +15,10 @@ export interface IProposalActionWithdrawTokenAsset {
      */
     symbol: string;
     /**
+     * Address of the token.
+     */
+    address?: string;
+    /**
      * URL of the token logo.
      */
     logo: string;