+++ packages/gov-ui-kit/src/modules/components/proposal/proposalActions/proposalActionsList/proposalActionWithdrawToken/proposalActionWithdrawToken.testUtils.ts
@@ -9,7 +9,7 @@ export const generateProposalActionWithdrawToken = (
     type: ProposalActionType.WITHDRAW_TOKEN,
     sender: { address: '0x742d35Cc6634C0532925a3b844Bc454e4438f44e' },
     receiver: { address: '0x3f5CE5FBFe3E9af3971dD833D26BA9b5C936F0bE' },
-    token: { name: 'Token name', symbol: 'TTT', logo: '', priceUsd: '1.00', decimals: 18 },
+    token: { name: 'Token name', symbol: 'TTT', logo: '', priceUsd: '1.00', decimals: 18, address: '0x1' },
     amount: '10000000',
     ...action,
 });